pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "Network agent automation and LangChain-based AI system"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.52.2",
    "paramiko",
    "pydantic===2.9.2",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "langgraph",
    "ipython",
    "requests",
    "typing-extensions",
    "deepagents",
    "langchain-mcp-adapters",
    "google-generativeai",
    "trulens-core",
    "trustcall",
    "mcp",
    "playwright>=1.57.0",
    "apscheduler",
    "chromadb",
    "pypdf2",
    "python-docx",
    "tqdm",
    "rich",
    "langchain-chroma>=1.1.0",
    "langchain-community",
    "langchain-text-splitters",
    "pypdf",
    "aio-pika>=9.0.0",
    "discord.py>=2.3.0",
]
